desktop/requirements.txt
argparse
requests
pyserial